+++ src/main/java/ru/skypro/homework/repository/CommentRepository.java
@@ -2,9 +2,9 @@ package ru.skypro.homework.repository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
-import ru.skypro.homework.Model.Comment;
+import ru.skypro.homework.model.entity.Comment;
 
 @Repository
-public interface CommentRepository extends JpaRepository<Comment, Long> {
+public interface CommentRepository extends JpaRepository<Comment, Integer> {
 
 }
