+++ src/main/java/ru/skypro/homework/repository/UserRepository.java
@@ -2,12 +2,12 @@ package ru.skypro.homework.repository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
-import ru.skypro.homework.Model.User;
+import ru.skypro.homework.model.entity.User;
 
 import java.util.Optional;
 
 @Repository
-public interface UserRepository extends JpaRepository<User, Long> {
+public interface UserRepository extends JpaRepository<User, Integer> {
 
    Optional<User> findUserByEmail(String email);
